feat: add clang to all images
Each project installs clang as a CI build step. Sometimes the package
install is throttled down to kB/s. Most of our rust projects use the
rdkafka crate which requires clang to build. It is worth the additional
image size of ~100MB and associated download time to eliminate huge
variance in CI build times when we are bandwidth throttled by the
package mirror