Fix Ruby install on 1.46-ruby-capnp image
This seems to have expected the `export PATH=...` command to persist
into the container's environment, but it seems to require an explicit
`ENV PATH=...` to accomplish that.

I'm also hardcoding to `/root` instead of `$HOME` because `$HOME`
expands to an empty string inside an `ENV` instruction (even though it's
correct in a `RUN` instruction).

Also adds the proper rbenv instantiation to /root/.bashrc.