Refactor rust-ubuntu image directory structure
And updated script/deploy to be able to deploy them. Now script/deploy
is the only deploy script.

I deleted `base-v1` because it appears to only be used in building the
rustc docker image. We tag images so infrequently I don't think it's
useful to factor that into its own image.

I deleted `base-v2` because it only appears to be used in the 1.44.1
image.